backend/requirements.txt
fastapi
uvicorn
transformers
torch
python-multipart
requests
gtts
pydub